move shared_example into specialized folder
