+++ spec/shared_examples/database_object.rb
